+++ MLEM.Ui/Elements/Element.cs
@@ -588,10 +588,12 @@ namespace MLEM.Ui.Elements {
                 index = this.children.Count;
             this.children.Insert(index, element);
             element.Parent = this;
-            element.AndChildren(e => e.AddedToUi(this.System, this.Root));
+            element.AndChildren(e => {
+                e.SetAreaDirty();
+                e.AddedToUi(this.System, this.Root);
+            });
             this.OnChildAdded?.Invoke(this, element);
             this.SetSortedChildrenDirty();
-            element.SetAreaDirty();
             return element;
         }