+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Additions
 Improvements
 - Allow scrolling panels to contain other scrolling panels
 - Allow dropdowns to have scrolling panels
+- Don't unnecessarily set areas dirty when removing a root element from the ui
 
 Fixes
 - Fixed panels updating their relevant children too much when the scroll bar is hidden